refactor: #668 domain layer에서 app layer로 변
